# Controllers/CNCMovementController.cs
using Microsoft.AspNetCore.Mvc;
using System.Threading.Tasks;
using HavenCNCServer.Models;

namespace HavenCNCServer.Controllers
{
    /// <summary>
    /// CNC Movement Control - Handles all movement, positioning, and fixture operations
    /// </summary>
    [ApiController]
    [Route("api/[controller]")]
    public class CNCMovementController : ControllerBase
    {
        #region Movement Control

        /// <summary>
        /// Set the movement type (relative or absolute)
        /// </summary>
        /// <param name="moveType">Movement type to set</param>
        /// <returns>Success response</returns>
        [HttpPost("SetMoveType")]
        public IActionResult SetMoveType([FromBody] MoveType moveType)
        {
            // TODO: Implement move type setting
            return Ok(new { message = $"Move type set to {moveType}", moveType });
        }

        /// <summary>
        /// Get the current movement type
        /// </summary>
        /// <returns>Current movement type</returns>
        [HttpGet("GetMoveType")]
        public IActionResult GetMoveType()
        {
            // TODO: Implement get move type
            return Ok(new { moveType = MoveType.Absolute });
        }

        /// <summary>
        /// Get current machine position
        /// </summary>
        /// <returns>Current position coordinates</returns>
        [HttpGet("GetCurrentPosition")]
        public IActionResult GetCurrentPosition()
        {
            // TODO: Implement get current position
            var position = new CNCPoint(0, 0, 0, 0);
            return Ok(position);
        }

        /// <summary>
        /// Move to specified coordinates
        /// </summary>
        /// <param name="request">Move request with coordinates and optional speeds</param>
        /// <returns>Success response</returns>
        [HttpPost("MoveTo")]
        public async Task<IActionResult> MoveTo([FromBody] MoveToRequest request)
        {
            // TODO: Implement move to functionality
            await Task.Delay(1);
            return Ok(new { message = $"Moving to {request.Point}", point = request.Point });
        }

        /// <summary>
        /// Move to coordinates until an IO event occurs
        /// </summary>
        /// <param name="request">Move request with coordinates, speeds, and IO event</param>
        /// <returns>Success response</returns>
        [HttpPost("MoveToUtil")]
        public async Task<IActionResult> MoveToUtil([FromBody] MoveToUntilRequest request)
        {
            // TODO: Implement move to until functionality
            await Task.Delay(1);
            return Ok(new { message = $"Moving to {request.Point} until {request.Event}", point = request.Point, ioEvent = request.Event });
        }

        /// <summary>
        /// Move in a direction until an IO event occurs
        /// </summary>
        /// <param name="request">Direction movement request</param>
        /// <returns>Success response</returns>
        [HttpPost("MoveDirectionUntil")]
        public async Task<IActionResult> MoveDirectionUntil([FromBody] MoveDirectionUntilRequest request)
        {
            // TODO: Implement directional move until functionality
            await Task.Delay(1);
            return Ok(new { message = $"Moving {request.Direction} until {request.Event}", direction = request.Direction, ioEvent = request.Event });
        }

        #endregion

        #region Fixture Management

        /// <summary>
        /// Change to a specific fixture
        /// </summary>
        /// <param name="fixtureNumber">Fixture number to change to</param>
        /// <returns>Success response</returns>
        [HttpPost("ChangeToFixture/{fixtureNumber}")]
        public async Task<IActionResult> ChangeToFixture(string fixtureNumber)
        {
            // TODO: Implement fixture change functionality
            await Task.Delay(1);
            return Ok(new { message = $"Changed to fixture {fixtureNumber}", fixtureNumber });
        }

        /// <summary>
        /// Set fixture to specific point
        /// </summary>
        /// <param name="request">Fixture setting request</param>
        /// <returns>Success response</returns>
        [HttpPost("SetFixtureToPoint")]
        public async Task<IActionResult> SetFixtureToPoint([FromBody] SetFixtureRequest request)
        {
            // TODO: Implement set fixture to point functionality
            await Task.Delay(1);
            return Ok(new { message = $"Set fixture {request.FixtureNumber} to {request.Point}", fixture = request.FixtureNumber, point = request.Point });
        }

        /// <summary>
        /// Set current fixture to specified point
        /// </summary>
        /// <param name="point">Point to set current fixture to</param>
        /// <returns>Success response</returns>
        [HttpPost("SetCurrentFixturePoint")]
        public async Task<IActionResult> SetCurrentFixturePoint([FromBody] CNCPoint point)
        {
            // TODO: Implement set current fixture point functionality
            await Task.Delay(1);
            return Ok(new { message = $"Set current fixture to {point}", point });
        }

        /// <summary>
        /// Get current fixture number
        /// </summary>
        /// <returns>Current fixture number</returns>
        [HttpGet("GetCurrentFixtureNumber")]
        public async Task<IActionResult> GetCurrentFixtureNumber()
        {
            // TODO: Implement get current fixture number
            await Task.Delay(1);
            return Ok(new { fixtureNumber = "G54" });
        }

        /// <summary>
        /// Save current Z position to current fixture
        /// </summary>
        /// <returns>Success response</returns>
        [HttpPost("SaveCurrentZToCurrentFixture")]
        public async Task<IActionResult> SaveCurrentZToCurrentFixture()
        {
            // TODO: Implement save Z to fixture functionality
            await Task.Delay(1);
            return Ok(new { message = "Saved current Z to current fixture" });
        }

        /// <summary>
        /// Save current XY position to current fixture
        /// </summary>
        /// <returns>Success response</returns>
        [HttpPost("SaveCurrentXYToCurrentFixture")]
        public async Task<IActionResult> SaveCurrentXYToCurrentFixture()
        {
            // TODO: Implement save XY to fixture functionality
            await Task.Delay(1);
            return Ok(new { message = "Saved current XY to current fixture" });
        }

        /// <summary>
        /// Save current XYZ position to current fixture
        /// </summary>
        /// <returns>Success response</returns>
        [HttpPost("SaveCurrentXYZToCurrentFixture")]
        public async Task<IActionResult> SaveCurrentXYZToCurrentFixture()
        {
            // TODO: Implement save XYZ to fixture functionality
            await Task.Delay(1);
            return Ok(new { message = "Saved current XYZ to current fixture" });
        }

        #endregion

        #region Feed Rate Control

        /// <summary>
        /// Get fast feed rate
        /// </summary>
        /// <returns>Current fast feed rate</returns>
        [HttpGet("GetFastFeedRate")]
        public async Task<IActionResult> GetFastFeedRate()
        {
            // TODO: Implement get fast feed rate
            await Task.Delay(1);
            return Ok(new { feedRate = 100.0 });
        }

        /// <summary>
        /// Set fast feed rate
        /// </summary>
        /// <param name="feedRate">Feed rate to set</param>
        /// <returns>Success response</returns>
        [HttpPost("SetFastFeedRate")]
        public async Task<IActionResult> SetFastFeedRate([FromBody] double feedRate)
        {
            // TODO: Implement set fast feed rate
            await Task.Delay(1);
            return Ok(new { message = $"Fast feed rate set to {feedRate}", feedRate });
        }

        /// <summary>
        /// Get normal feed rate
        /// </summary>
        /// <returns>Current normal feed rate</returns>
        [HttpGet("GetNormalFeedRate")]
        public async Task<IActionResult> GetNormalFeedRate()
        {
            // TODO: Implement get normal feed rate
            await Task.Delay(1);
            return Ok(new { feedRate = 50.0 });
        }

        /// <summary>
        /// Set normal feed rate
        /// </summary>
        /// <param name="feedRate">Feed rate to set</param>
        /// <returns>Success response</returns>
        [HttpPost("SetNormalFeedRate")]
        public async Task<IActionResult> SetNormalFeedRate([FromBody] double feedRate)
        {
            // TODO: Implement set normal feed rate
            await Task.Delay(1);
            return Ok(new { message = $"Normal feed rate set to {feedRate}", feedRate });
        }

        /// <summary>
        /// Adjust normal feed rate by factor
        /// </summary>
        /// <param name="factor">Adjustment factor (-200 to +200, where 0 is no change, 100 is double, -100 is half)</param>
        /// <returns>Success response</returns>
        [HttpPost("AdjustNormalFeedRate")]
        public async Task<IActionResult> AdjustNormalFeedRate([FromBody] double factor)
        {
            // TODO: Implement adjust normal feed rate
            await Task.Delay(1);
            return Ok(new { message = $"Normal feed rate adjusted by factor {factor}", factor });
        }

        /// <summary>
        /// Get current normal feed rate adjustment factor
        /// </summary>
        /// <returns>Current adjustment factor</returns>
        [HttpGet("GetCurrentNormalFeedRateFactor")]
        public async Task<IActionResult> GetCurrentNormalFeedRateFactor()
        {
            // TODO: Implement get current normal feed rate factor
            await Task.Delay(1);
            return Ok(new { factor = 0.0 });
        }

        /// <summary>
        /// Reset normal feed rate factor to default
        /// </summary>
        /// <returns>Success response</returns>
        [HttpPost("ResetNormalFeedRateFactor")]
        public async Task<IActionResult> ResetNormalFeedRateFactor()
        {
            // TODO: Implement reset normal feed rate factor
            await Task.Delay(1);
            return Ok(new { message = "Normal feed rate factor reset to default" });
        }

        /// <summary>
        /// Adjust fast feed rate by factor
        /// </summary>
        /// <param name="factor">Adjustment factor (-200 to +200, where 0 is no change, 100 is double, -100 is half)</param>
        /// <returns>Success response</returns>
        [HttpPost("AdjustFastFeedRate")]
        public async Task<IActionResult> AdjustFastFeedRate([FromBody] double factor)
        {
            // TODO: Implement adjust fast feed rate
            await Task.Delay(1);
            return Ok(new { message = $"Fast feed rate adjusted by factor {factor}", factor });
        }

        /// <summary>
        /// Get current fast feed rate adjustment factor
        /// </summary>
        /// <returns>Current adjustment factor</returns>
        [HttpGet("GetCurrentFastFeedRateFactor")]
        public async Task<IActionResult> GetCurrentFastFeedRateFactor()
        {
            // TODO: Implement get current fast feed rate factor
            await Task.Delay(1);
            return Ok(new { factor = 0.0 });
        }

        /// <summary>
        /// Reset fast feed rate factor to default
        /// </summary>
        /// <returns>Success response</returns>
        [HttpPost("ResetFastFeedRateFactor")]
        public async Task<IActionResult> ResetFastFeedRateFactor()
        {
            // TODO: Implement reset fast feed rate factor
            await Task.Delay(1);
            return Ok(new { message = "Fast feed rate factor reset to default" });
        }

        #endregion
    }
}