# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "HavenCNC Server API",
    "description": "REST API for CNC server management hosted within WinForms application",
    "contact": {
      "name": "HavenCNC Server Team"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/CNCMovement/SetMoveType": {
      "post": {
        "tags": [
          "CNCMovement"
        ],
        "summary": "Set the movement type (relative or absolute)",
        "requestBody": {
          "description": "Movement type to set",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MoveType"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/CNCMovement/MoveTo": {
      "post": {
        "tags": [
          "CNCMovement"
        ],
        "summary": "Move to specified coordinates",
        "requestBody": {
          "description": "Move request with coordinates and optional speeds",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveToRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveToRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MoveToRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/CNCMovement/MoveToUtil": {
      "post": {
        "tags": [
          "CNCMovement"
        ],
        "summary": "Move to coordinates until an IO event occurs",
        "requestBody": {
          "description": "Move request with coordinates, speeds, and IO event",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveToUntilRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveToUntilRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MoveToUntilRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/CNCMovement/MoveDirectionUntil": {
      "post": {
        "tags": [
          "CNCMovement"
        ],
        "summary": "Move in a direction until an IO event occurs",
        "requestBody": {
          "description": "Direction movement request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveDirectionUntilRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveDirectionUntilRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MoveDirectionUntilRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/CNCMovement/ChangeToFixture/{fixtureNumber}": {
      "post": {
        "tags": [
          "CNCMovement"
        ],
        "summary": "Change to a specific fixture",
        "parameters": [
          {
            "name": "fixtureNumber",
            "in": "path",
            "description": "Fixture number to change to",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/CNCMovement/SetFixtureToPoint": {
      "post": {
        "tags": [
          "CNCMovement"
        ],
        "summary": "Set fixture to specific point",
        "requestBody": {
          "description": "Fixture setting request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetFixtureRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetFixtureRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetFixtureRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/CNCMovement/SetCurrentFixturePoint": {
      "post": {
        "tags": [
          "CNCMovement"
        ],
        "summary": "Set current fixture to specified point",
        "requestBody": {
          "description": "Point to set current fixture to",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CNCPoint"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CNCPoint"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CNCPoint"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/CNCMovement/SetFastFeedRate": {
      "post": {
        "tags": [
          "CNCMovement"
        ],
        "summary": "Set fast feed rate",
        "requestBody": {
          "description": "Feed rate to set",
          "content": {
            "application/json": {
              "schema": {
                "type": "number",
                "format": "double"
              }
            },
            "text/json": {
              "schema": {
                "type": "number",
                "format": "double"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "number",
                "format": "double"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/CNCMovement/SetNormalFeedRate": {
      "post": {
        "tags": [
          "CNCMovement"
        ],
        "summary": "Set normal feed rate",
        "requestBody": {
          "description": "Feed rate to set",
          "content": {
            "application/json": {
              "schema": {
                "type": "number",
                "format": "double"
              }
            },
            "text/json": {
              "schema": {
                "type": "number",
                "format": "double"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "number",
                "format": "double"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/CNCMovement/AdjustNormalFeedRate": {
      "post": {
        "tags": [
          "CNCMovement"
        ],
        "summary": "Adjust normal feed rate by factor",
        "requestBody": {
          "description": "Adjustment factor (-200 to +200, where 0 is no change, 100 is double, -100 is half)",
          "content": {
            "application/json": {
              "schema": {
                "type": "number",
                "format": "double"
              }
            },
            "text/json": {
              "schema": {
                "type": "number",
                "format": "double"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "number",
                "format": "double"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/CNCMovement/AdjustFastFeedRate": {
      "post": {
        "tags": [
          "CNCMovement"
        ],
        "summary": "Adjust fast feed rate by factor",
        "requestBody": {
          "description": "Adjustment factor (-200 to +200, where 0 is no change, 100 is double, -100 is half)",
          "content": {
            "application/json": {
              "schema": {
                "type": "number",
                "format": "double"
              }
            },
            "text/json": {
              "schema": {
                "type": "number",
                "format": "double"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "number",
                "format": "double"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CNCPoint": {
        "type": "object",
        "properties": {
          "x": {
            "type": "number",
            "description": "X-axis coordinate",
            "format": "double"
          },
          "y": {
            "type": "number",
            "description": "Y-axis coordinate",
            "format": "double"
          },
          "z": {
            "type": "number",
            "description": "Z-axis coordinate",
            "format": "double"
          },
          "a": {
            "type": "number",
            "description": "A-axis coordinate (rotational axis)",
            "format": "double"
          }
        },
        "additionalProperties": false,
        "description": "Represents a point in CNC machine coordinates with X, Y, Z, and A axis values"
      },
...
      "IOEvent": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "description": "IO events that can be monitored or triggered",
        "format": "int32"
      },
...
      "MoveDirection": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "description": "Directions for machine movement",
        "format": "int32"
      },
      "MoveDirectionUntilRequest": {
        "type": "object",
        "properties": {
          "direction": {
            "$ref": "#/components/schemas/MoveDirection"
          },
          "event": {
            "$ref": "#/components/schemas/IOEvent"
          },
          "maxDistance": {
            "type": "number",
            "description": "Maximum distance to move (optional)",
            "format": "double",
            "nullable": true
          },
          "speed": {
            "type": "number",
            "description": "Movement speed (optional)",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Request model for directional movement until an event"
      },
      "MoveToRequest": {
        "type": "object",
        "properties": {
          "point": {
            "$ref": "#/components/schemas/CNCPoint"
          },
          "xySpeed": {
            "type": "number",
            "description": "XY movement speed (optional)",
            "format": "double",
            "nullable": true
          },
          "zSpeed": {
            "type": "number",
            "description": "Z movement speed (optional)",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Request model for moving to a specific point"
      },
      "MoveToUntilRequest": {
        "type": "object",
        "properties": {
          "point": {
            "$ref": "#/components/schemas/CNCPoint"
          },
          "xySpeed": {
            "type": "number",
            "description": "XY movement speed (optional)",
            "format": "double",
            "nullable": true
          },
          "zSpeed": {
            "type": "number",
            "description": "Z movement speed (optional)",
            "format": "double",
            "nullable": true
          },
          "event": {
            "$ref": "#/components/schemas/IOEvent"
          }
        },
        "additionalProperties": false,
        "description": "Request model for moving until an IO event occurs"
      },
      "MoveType": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "description": "Types of movement commands",
        "format": "int32"
      },
...
      "SetFixtureRequest": {
        "type": "object",
        "properties": {
          "fixtureNumber": {
            "type": "string",
            "description": "Fixture number",
            "nullable": true
          },
          "point": {
            "$ref": "#/components/schemas/CNCPoint"
          }
        },
        "additionalProperties": false,
        "description": "Request model for setting fixture"
      },
...
    }
  }
}